requirements.txt
streamlit
mysql-connector-python
onnx
onnxruntime
transformers
accelerate
evaluate
nltk
watchdog
